Try to disentangle vision and audio data code
