+++ conduit/data/__init__.py
@@ -1,4 +1,2 @@
 from .constants import *
-from .datamodules import *
-from .datasets import *
 from .structures import *
