+++ conduit/data/datamodules/__init__.py
@@ -1,3 +1,1 @@
-from .audio import *
 from .base import *
-from .vision import *
