+++ conduit/data/datasets/__init__.py
@@ -1,5 +1,3 @@
-from .audio import *
 from .base import *
 from .utils import *
-from .vision import *
 from .wrappers import *
