+++ conduit/data/datasets/tabular/__init__.py
@@ -1,1 +1,2 @@
 from .base import *
+from .dummy import *
