+++ conduit/data/datasets/vision/isic.py
@@ -18,9 +18,10 @@ from torch import Tensor
 from tqdm import tqdm
 from typing_extensions import TypeAlias
 
-from conduit.data.datasets.utils import ImageTform
-from conduit.data.datasets.vision.base import CdtVisionDataset
 from conduit.data.structures import TernarySample
+
+from .base import CdtVisionDataset
+from .utils import ImageTform
 
 __all__ = ["IsicAttr", "ISIC"]
