+++ conduit/data/datasets/vision/iwildcam.py
@@ -7,9 +7,11 @@ import torch
 from torch import Tensor
 from typing_extensions import TypeAlias
 
-from conduit.data.datasets.utils import ImageTform, UrlFileInfo, download_from_url
-from conduit.data.datasets.vision.base import CdtVisionDataset
+from conduit.data.datasets.utils import UrlFileInfo, download_from_url
 from conduit.data.structures import SubgroupSample, TernarySample
+
+from .base import CdtVisionDataset
+from .utils import ImageTform
 
 __all__ = ["IWildCam", "IWildCamSplit", "IWildCamUnlabeled"]
