+++ pyproject.toml
@@ -30,12 +30,12 @@ scikit-learn = "^1.2.0"
 typing-extensions = ">=4.4.0"
 # optional dependencies
 albumentations = {version = "^1.0.0", optional = true}
-ethicml = {verion = ">=1.0.2", extras = ["data"], optional = true}
+ethicml = {version = ">=1.0.2", extras = ["data"], optional = true}
 gdown = {version = "^3.13.0", optional = true}
-hydra-core = {verion = "^1.1.1", optional = true}
+hydra-core = {version = "^1.1.1", optional = true}
 kaggle = {version = "^1.5.12", optional = true}
 opencv-python = {version = "^4.5.3", optional = true}
-rich = {verion = "^12.5.1", optional = true}
+rich = {version = "^12.5.1", optional = true}
 soundfile = {version = "*", markers = "platform_system == 'Windows'", optional = true}
 sox = {version = "*", markers = "platform_system == 'Linux' or platform_system == 'macOS'", optional = true}
 
@@ -110,7 +110,7 @@ extra_standard_library = ["dataclasses", "__future__", "typing_extensions"]
 line_length = 88
 profile = "black"
 force_sort_within_sections = "True"
-classes = ["MISSING", "NICO", "PACS", "SSRP"]
+classes = ["MISSING", "NICO", "PACS", "SSRP", "ISIC"]
 
 [tool.pyright]
 include = ["conduit"]
